requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
multidict==6.6.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
py-cord==2.6.1
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
audioop-lts
pymongo
dnspython
certifi